Code for the Kokkos/cuda examples using ADIOS
